Implement Rename
Possible rename/mv scenarios are handled as follows:

    OldPath      NewPathExists       NewPath     Result
    -------------------------------------------------------
    File         No                  N/A         Rename
    File         Yes                 File        Overwrite
    File         Yes                 Dir         Error!
    Dir          No                  N/A         Rename
    Dir          Yes                 File        Error!
    Dir          Yes                 Dir         Overwrite if empty, else Error!

For the last case, say we have the following tree structure, a and a'
are the same name but are distinguished to be able to talk about them.

    ├── a
    └── b
        └── a'
           [...]

If we try 'mv a b', it should fail if a' is non-empty. If a' is empty, we
should override to get the following:

    └── b
        └── a